requirements.txt
python-telegram-bot==21.6
aiohttp==3.10.5
python-dotenv==1.0.1
qbittorrent-api==2024.2.59
deep-translator==1.11.4
groq==0.11.0
